Create generate default function
